projects/auth/src/requirements.txt
django==4.1.4
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5